Allow api commands to set curl options
The specific use case is when headers need to be set to, for instance,
identify acceptable content types (application/json).  To use, create a
bash array and set to the gls_basic_api_curl_opts variable.  Using an
array vs direct string interpolation makes it less error-prone when
setting values that have whitespace in them.